+++ rtca/backend/websocket.go
@@ -3,7 +3,6 @@ package main
 import (
 	"log"
 	"net/http"
-
 	"github.com/gorilla/websocket"
 )
 
@@ -56,14 +55,4 @@ func handleMessages() {
 			}
 		}
 	}
-}
-
-func main() {
-	mainRoutine() //testing db
-	
-	//Invokes Web Socket connection
-	http.HandleFunc("/ws", handleConnections) //forever looping
-	go handleMessages()                       	 //forever looping
-	log.Println("WebSocket server started at ws://localhost:8080/ws")
-	log.Fatal(http.ListenAndServe(":8080", nil)) //loops until error
 }
